requirements.txt
google-genai>=1.70.0
python-dotenv>=1.0.0
Flask>=3.0.0
